+++ sertit/__init__.py
@@ -20,4 +20,4 @@
 .. include:: ../README.md
 """
 
-__version__ = "1.3.7"
+__version__ = "1.3.8"